Add cost calculation

+++ models/open_ai_model.py
@@ -1,5 +1,7 @@
+from decimal import Decimal
+
 class OpenAIModel:
-    def __init__(self, model_id: str, input_token_price: float, output_token_price: float) -> None:
+    def __init__(self, model_id: str, input_token_price: Decimal, output_token_price: Decimal) -> None:
         self.model_id = model_id
         self.input_token_price = input_token_price
         self.output_token_price = output_token_price
